* Added -Wall to gcc compilations * Added support for run_async() and wait(), extending usability of ThreadPool class
git-svn-id: https://svn.code.sf.net/p/fract/code@268 4b73aca3-9014-0410-b8ed-9bdd1c1cebd0

+++ cxxptl/tester.cpp
@@ -71,7 +71,7 @@ int main(void)
 			sum += dowork.data[i*100];
 		}
 		printf("pi = %.9lf\n", sqrt(sum * 6.0));
-		printf("Result produced in %u sec\n", xt);
+		printf("Result produced in %u sec\n", (unsigned)xt);
 	}
 	return 0;
 }
